Create FillXNumberOfRowsConstraintStrategy
